added configuration depend on interface
