+++ src/test/java/org/selenium/demo/pages/CityPage.java
@@ -1,4 +1,4 @@
-package org.selenium.demo;
+package org.selenium.demo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -6,24 +6,25 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
 /**
- * Created by web_systems on 17.03.2017.
+ * Created by web_systems on 29.03.2017.
  */
-public class NamePage {
+public class CityPage {
     private WebDriver driver;
-    public NamePage(WebDriver driver) {
+    public CityPage(WebDriver driver) {
         PageFactory.initElements(driver, this);
         this.driver = driver;
     }
-    @FindBy(id = "q1")
-    private WebElement nametext;
+    @FindBy(id = "q6")
+    private WebElement citytext;
 
     @FindBy(xpath = "/html/body/div[5]/div/div/div/div/div/div[2]/div[1]/button")
-    private WebElement nameclick;
+    private WebElement cityclick;
 
-    public void writeName(String text) {
-        nametext.sendKeys(text);
+    public void writeCity(String text){
+        citytext.sendKeys(text);
+
     }
-    public void clickName () {
-        nameclick.click();
+    public void clickCity(){
+        cityclick.click();
     }
 }
