+++ src/test/java/org/selenium/demo/pages/ComentPage.java
@@ -1,4 +1,4 @@
-package org.selenium.demo;
+package org.selenium.demo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -6,25 +6,24 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
 /**
- * Created by web_systems on 29.03.2017.
+ * Created by web_systems on 17.03.2017.
  */
-public class CityPage {
+public class ComentPage {
     private WebDriver driver;
-    public CityPage(WebDriver driver) {
+    public ComentPage(WebDriver driver) {
         PageFactory.initElements(driver, this);
         this.driver = driver;
     }
-    @FindBy(id = "q6")
-    private WebElement citytext;
+    @FindBy(id = "q4")
+    private WebElement comenttext;
 
     @FindBy(xpath = "/html/body/div[5]/div/div/div/div/div/div[2]/div[1]/button")
-    private WebElement cityclick;
+    private WebElement comentclick;
 
-    public void writeCity(String text){
-        citytext.sendKeys(text);
-
+    public void writeComent(String text){
+        comenttext.sendKeys(text);
     }
-    public void clickCity(){
-        cityclick.click();
+    public void  clickComent(){
+        comentclick.click();
     }
 }
