+++ src/test/java/org/selenium/demo/pages/EmailPage.java
@@ -1,4 +1,4 @@
-package org.selenium.demo;
+package org.selenium.demo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -8,22 +8,23 @@ import org.openqa.selenium.support.PageFactory;
 /**
  * Created by web_systems on 17.03.2017.
  */
-public class PriorityPage {
+public class EmailPage {
     private WebDriver driver;
-    public PriorityPage(WebDriver driver) {
+    public EmailPage(WebDriver driver) {
         PageFactory.initElements(driver, this);
         this.driver = driver;
     }
-    @FindBy(xpath = "//*[@id=\"Слой_1\"]")
-    private WebElement priorityclick;
+    @FindBy(id="q2")
+    private WebElement emailtext;
 
     @FindBy(xpath = "/html/body/div[5]/div/div/div/div/div/div[2]/div[1]/button")
-    private WebElement bpriorityclick;
+    private WebElement emailclick;
 
-    public void clickPriority(){
-        priorityclick.click();
+    public void writeEmail(String text){
+        emailtext.sendKeys(text);
+
     }
-    public void clickBpiority(){
-        bpriorityclick.click();
+    public void clickEmail(){
+        emailclick.click();
     }
 }
