+++ src/test/java/org/selenium/demo/pages/FinishPage.java
@@ -1,4 +1,4 @@
-package org.selenium.demo;
+package org.selenium.demo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
