+++ src/test/java/org/selenium/demo/pages/NamePage.java
@@ -1,4 +1,4 @@
-package org.selenium.demo;
+package org.selenium.demo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -8,22 +8,22 @@ import org.openqa.selenium.support.PageFactory;
 /**
  * Created by web_systems on 17.03.2017.
  */
-public class BudgetePage {
+public class NamePage {
     private WebDriver driver;
-    public BudgetePage(WebDriver driver) {
+    public NamePage(WebDriver driver) {
         PageFactory.initElements(driver, this);
         this.driver = driver;
     }
-    @FindBy(id = "q5")
-    private WebElement budgetwrite;
+    @FindBy(id = "q1")
+    private WebElement nametext;
 
     @FindBy(xpath = "/html/body/div[5]/div/div/div/div/div/div[2]/div[1]/button")
-    private WebElement budgetclick;
+    private WebElement nameclick;
 
-    public void writeBudget(String text){
-        budgetwrite.sendKeys(text);
+    public void writeName(String text) {
+        nametext.sendKeys(text);
     }
-    public void clickBudget(){
-        budgetclick.click();
+    public void clickName () {
+        nameclick.click();
     }
 }
