+++ src/test/java/org/selenium/demo/pages/Page.java
@@ -1,0 +1,7 @@
+package org.selenium.demo.pages;
+
+public interface Page {
+
+    void click();
+    void writeText(String text);
+}
